python/requirements.txt
pandas==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.0
matplotlib==3.8.0